Store counted strings in StringPool (#16)
The `StringPool` storage currently stores string items as a sequence of NUL-terminated C-strings, identified by a `StringId` type which represents the starting offset for the value.

The main disadvantage of this is that it cannot accommodate binary data. This PR updates the pool to use counted strings instead. It adds `CountedString` to manage instances, with the length encoded in a prefix of one or two bytes.

+++ src/include/ConfigDB/Object.h
@@ -152,7 +152,7 @@ protected:
 
 	String getString(StringId id) const;
 
-	StringId getStringId(const char* value, size_t valueLength);
+	StringId getStringId(const char* value, uint16_t valueLength);
 
 	StringId getStringId(const String& value)
 	{
